+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 [0.9.0]: https://github.com/mhassan1/yarn-plugin-licenses/compare/v0.8.1...v0.9.0
 
 - Add support for `repository` strings
+- Normalize `shortcut`-type `repository` URLs
 
 ## [0.8.1] - 2022-04-13
 [0.8.1]: https://github.com/mhassan1/yarn-plugin-licenses/compare/v0.8.0...v0.8.1